refactor(plugin): share asset emission helpers
