+++ crates/rolldown_plugin_asset_module/Cargo.toml
@@ -13,11 +13,8 @@ test = false
 
 [dependencies]
 anyhow = { workspace = true }
-memchr = { workspace = true }
 rolldown_common = { workspace = true }
 rolldown_plugin = { workspace = true }
-rolldown_std_utils = { workspace = true }
+rolldown_plugin_utils = { workspace = true }
 rolldown_utils = { workspace = true }
 rustc-hash = { workspace = true }
-string_wizard = { workspace = true }
-tokio = { workspace = true, features = ["fs"] }
