+++ crates/rolldown_plugin_asset_module/src/lib.rs
@@ -1,16 +1,13 @@
-use std::{borrow::Cow, path::Path, sync::Arc};
+use std::{borrow::Cow, path::Path};
 
-use memchr::memmem;
-use rolldown_common::{EmittedAsset, ModuleType, StrOrBytes, side_effects::HookSideEffects};
+use rolldown_common::{ModuleType, side_effects::HookSideEffects};
 use rolldown_plugin::{
-  HookLoadArgs, HookLoadOutput, HookLoadReturn, HookRenderChunkArgs, HookRenderChunkOutput,
-  HookRenderChunkReturn, HookTransformOutputMap, HookUsage, Plugin, PluginHookMeta, PluginOrder,
-  SharedLoadPluginContext,
+  HookLoadArgs, HookLoadOutput, HookLoadReturn, HookRenderChunkArgs, HookRenderChunkReturn,
+  HookUsage, Plugin, PluginHookMeta, PluginOrder, SharedLoadPluginContext,
 };
-use rolldown_std_utils::relative_path_as_js_specifier;
+use rolldown_plugin_utils::{emit_asset, rewrite_emitted_asset_references};
 use rolldown_utils::url::clean_url;
 use rustc_hash::FxHashSet;
-use string_wizard::{MagicString, SourceMapOptions};
 
 const PREFIX: &str = "__ROLLDOWN_ASSET__#";
 
@@ -64,61 +61,7 @@ impl Plugin for AssetModulePlugin {
     ctx: &rolldown_plugin::PluginContext,
     args: &HookRenderChunkArgs<'_>,
   ) -> HookRenderChunkReturn {
-    // Quick bail: if the code doesn't contain our prefix, nothing to do
-    if !args.code.contains(PREFIX) {
-      return Ok(None);
-    }
-
-    let chunk_filename = &args.chunk.filename;
-    let code = args.code.as_str();
-    let mut magic_string = MagicString::new(code);
-    let mut changed = false;
-
-    // Use memchr for SIMD-accelerated substring search
-    let finder = memmem::find_iter(code.as_bytes(), PREFIX.as_bytes());
-
-    for abs_pos in finder {
-      let after_prefix = abs_pos + PREFIX.len();
-
-      // Extract ref_id: scan until we hit a quote (", ') or end of string
-      let rest = &code[after_prefix..];
-      let ref_end = rest.find(['"', '\'']).unwrap_or(rest.len());
-      let ref_id = &rest[..ref_end];
-
-      if ref_id.is_empty() {
-        continue;
-      }
-
-      // Resolve the asset filename
-      let asset_filename = match ctx.get_file_name(ref_id) {
-        Ok(name) => name,
-        Err(_) => continue,
-      };
-
-      // Compute relative path from chunk to asset
-      let relative = compute_relative_path(chunk_filename, &asset_filename);
-
-      let end = after_prefix + ref_end;
-      #[expect(clippy::cast_possible_truncation)]
-      if magic_string.update(abs_pos as u32, end as u32, relative).is_ok() {
-        changed = true;
-      }
-    }
-
-    if changed {
-      Ok(Some(HookRenderChunkOutput {
-        code: magic_string.to_string(),
-        map: HookTransformOutputMap::from_if_enabled(args.options.sourcemap.is_some(), || {
-          magic_string.source_map(SourceMapOptions {
-            hires: string_wizard::Hires::Boundary,
-            include_content: false,
-            source: Arc::from(args.chunk.filename.as_str()),
-          })
-        }),
-      }))
-    } else {
-      Ok(None)
-    }
+    Ok(rewrite_emitted_asset_references(ctx, args, PREFIX))
   }
 }
 
@@ -142,36 +85,14 @@ impl AssetModulePlugin {
       return Ok(None);
     }
 
-    let path = Path::new(clean_id);
-
-    // Read file as binary (use cleaned path for filesystem access)
-    let bytes = tokio::fs::read(clean_id)
-      .await
-      .map_err(|e| anyhow::anyhow!("Failed to read asset module {clean_id}: {e}"))?;
-
-    // Derive name from the cleaned file path
-    let file_name = path.file_name().and_then(|n| n.to_str()).unwrap_or("asset").to_string();
-
-    // Use relative path for original_file_name to avoid leaking absolute paths
-    let original_file_name =
-      path.strip_prefix(ctx.cwd()).unwrap_or(path).to_string_lossy().into_owned();
-
-    // Emit the file as an asset
-    let reference_id = ctx
-      .emit_file_async(EmittedAsset {
-        name: Some(file_name),
-        original_file_name: Some(original_file_name),
-        source: StrOrBytes::Bytes(bytes),
-        ..Default::default()
-      })
-      .await?;
-
-    // Associate this module with the emitted file so the `new URL()` finalizer
-    // can look up the asset filename by module ID (use original args.id for mapping)
-    ctx.associate_module_with_file_ref(args.id, &reference_id);
-
-    // Add watch file for watch mode (use cleaned path)
-    ctx.add_watch_file(clean_id);
+    let reference_id = emit_asset(
+      &ctx,
+      clean_id,
+      |e| anyhow::anyhow!("Failed to read asset module {clean_id}: {e}"),
+      |reference_id| ctx.associate_module_with_file_ref(args.id, reference_id),
+      |clean_id| ctx.add_watch_file(clean_id),
+    )
+    .await?;
 
     // Return JS code that exports the asset placeholder via CJS.
     // Using `module.exports` ensures `require()` returns the string directly.
@@ -197,12 +118,4 @@ impl AssetModulePlugin {
       .and_then(|e| e.to_str())
       .is_some_and(|ext| self.asset_extensions.contains(ext))
   }
-}
-
-/// Compute the relative path from a chunk file to an asset file,
-/// ensuring it starts with "./" for relative paths.
-fn compute_relative_path(chunk_filename: &str, asset_filename: &str) -> String {
-  let chunk_dir = Path::new(chunk_filename).parent().unwrap_or(Path::new(""));
-
-  relative_path_as_js_specifier(asset_filename, chunk_dir)
 }
