+++ crates/rolldown_plugin_copy_module/src/lib.rs
@@ -1,17 +1,14 @@
-use std::{borrow::Cow, path::Path, sync::Arc};
+use std::{borrow::Cow, path::Path};
 
 use arcstr::ArcStr;
-use memchr::memmem;
-use rolldown_common::{EmittedAsset, ModuleType, ResolvedExternal, StrOrBytes};
+use rolldown_common::{ModuleType, ResolvedExternal};
 use rolldown_plugin::{
-  HookRenderChunkArgs, HookRenderChunkOutput, HookRenderChunkReturn, HookResolveIdArgs,
-  HookResolveIdOutput, HookResolveIdReturn, HookTransformOutputMap, HookUsage, Plugin,
-  PluginContext, PluginHookMeta, PluginOrder,
+  HookRenderChunkArgs, HookRenderChunkReturn, HookResolveIdArgs, HookResolveIdOutput,
+  HookResolveIdReturn, HookUsage, Plugin, PluginContext, PluginHookMeta, PluginOrder,
 };
-use rolldown_std_utils::relative_path_as_js_specifier;
+use rolldown_plugin_utils::{emit_asset, rewrite_emitted_asset_references};
 use rolldown_utils::url::clean_url;
 use rustc_hash::FxHashSet;
-use string_wizard::{MagicString, SourceMapOptions};
 
 const PREFIX: &str = "__ROLLDOWN_COPY_MODULE__#";
 
@@ -94,31 +91,14 @@ impl Plugin for CopyModulePlugin {
       return Ok(None);
     }
 
-    // Read the file bytes asynchronously to avoid blocking the tokio worker thread
-    let bytes = tokio::fs::read(clean_id)
-      .await
-      .map_err(|e| anyhow::anyhow!("Failed to read copy module {}: {e}", resolved_id.id))?;
-
-    // Derive a name from the file path
-    let file_name =
-      resolved_path.file_name().and_then(|n| n.to_str()).unwrap_or("asset").to_string();
-
-    // Use relative path for original_file_name to avoid leaking absolute paths into output
-    let original_file_name =
-      resolved_path.strip_prefix(ctx.cwd()).unwrap_or(resolved_path).to_string_lossy().into_owned();
-
-    // Emit the file as an asset
-    let reference_id = ctx
-      .emit_file_async(EmittedAsset {
-        name: Some(file_name),
-        original_file_name: Some(original_file_name),
-        source: StrOrBytes::Bytes(bytes),
-        ..Default::default()
-      })
-      .await?;
-
-    // Add watch file for watch mode (use the clean absolute path)
-    ctx.add_watch_file(clean_id);
+    let reference_id = emit_asset(
+      ctx,
+      clean_id,
+      |e| anyhow::anyhow!("Failed to read copy module {}: {e}", resolved_id.id),
+      |_| {},
+      |clean_id| ctx.add_watch_file(clean_id),
+    )
+    .await?;
 
     // Return a prefixed external ID — the prefix will be rewritten in render_chunk
     let placeholder_id: ArcStr = format!("{PREFIX}{reference_id}").into();
@@ -141,68 +121,6 @@ impl Plugin for CopyModulePlugin {
     ctx: &PluginContext,
     args: &HookRenderChunkArgs<'_>,
   ) -> HookRenderChunkReturn {
-    // Quick bail: if the code doesn't contain our prefix, nothing to do
-    if !args.code.contains(PREFIX) {
-      return Ok(None);
-    }
-
-    let chunk_filename = &args.chunk.filename;
-    let code = args.code.as_str();
-    let mut magic_string = MagicString::new(code);
-    let mut changed = false;
-
-    // Use memchr for SIMD-accelerated substring search
-    let finder = memmem::find_iter(code.as_bytes(), PREFIX.as_bytes());
-
-    for abs_pos in finder {
-      let after_prefix = abs_pos + PREFIX.len();
-
-      // Extract ref_id: scan until we hit a quote (", ') or end of string
-      let rest = &code[after_prefix..];
-      let ref_end = rest.find(['"', '\'']).unwrap_or(rest.len());
-      let ref_id = &rest[..ref_end];
-
-      if ref_id.is_empty() {
-        continue;
-      }
-
-      // Resolve the asset filename
-      let asset_filename = match ctx.get_file_name(ref_id) {
-        Ok(name) => name,
-        Err(_) => continue,
-      };
-
-      // Compute relative path from chunk to asset
-      let relative = compute_relative_path(chunk_filename, &asset_filename);
-
-      let end = after_prefix + ref_end;
-      #[expect(clippy::cast_possible_truncation)]
-      if magic_string.update(abs_pos as u32, end as u32, relative).is_ok() {
-        changed = true;
-      }
-    }
-
-    if changed {
-      Ok(Some(HookRenderChunkOutput {
-        code: magic_string.to_string(),
-        map: HookTransformOutputMap::from_if_enabled(args.options.sourcemap.is_some(), || {
-          magic_string.source_map(SourceMapOptions {
-            hires: string_wizard::Hires::Boundary,
-            include_content: false,
-            source: Arc::from(args.chunk.filename.as_str()),
-          })
-        }),
-      }))
-    } else {
-      Ok(None)
-    }
+    Ok(rewrite_emitted_asset_references(ctx, args, PREFIX))
   }
-}
-
-/// Compute the relative path from a chunk file to an asset file,
-/// ensuring it starts with "./" for relative paths.
-fn compute_relative_path(chunk_filename: &str, asset_filename: &str) -> String {
-  let chunk_dir = Path::new(chunk_filename).parent().unwrap_or(Path::new(""));
-
-  relative_path_as_js_specifier(asset_filename, chunk_dir)
 }
