+++ crates/rolldown_plugin_utils/Cargo.toml
@@ -19,9 +19,13 @@ workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
+arcstr = { workspace = true }
 memchr = { workspace = true }
 oxc = { workspace = true }
 rolldown_common = { workspace = true }
 rolldown_plugin = { workspace = true }
+rolldown_std_utils = { workspace = true }
 rolldown_utils = { workspace = true }
 serde_json = { workspace = true }
+string_wizard = { workspace = true }
+tokio = { workspace = true, features = ["fs"] }
