+++ crates/rolldown_plugin_utils/src/lib.rs
@@ -1,3 +1,4 @@
+mod asset_emission;
 mod data_to_esm;
 mod is_special_query;
 mod parse_program;
@@ -6,6 +7,7 @@ mod to_string_literal;
 
 pub mod constants;
 
+pub use asset_emission::{emit_asset, rewrite_emitted_asset_references};
 pub use data_to_esm::data_to_esm;
 pub use is_special_query::is_special_query;
 pub use parse_program::parse_program;
